Add board timer reset for long shutdowns
After triggering shutdown process the x708 board
is waiting for ~50 seconds for Pi to shutdown.
If Pi shutdown doesn't take place in this range
then the x708 doesn't cut off the power to the
Pi afterwards.
This commit resolves the issue by triggering the
shutdown process on x708 in a background loop with
a manually tested timeout.
Removed WIP label from the README.